+++ client/js/modals/admin-users-delete.js
@@ -1,11 +1,11 @@
-/* global $, Vue */
+/* global $, Vue, usrData, alerts */
 
 // Vue Delete User instance
 
 let vueDeleteUser = new Vue({
   el: '#modal-admin-users-delete',
   data: {
-
+    loading: false
   },
   methods: {
     open: (ev) => {
@@ -15,7 +15,18 @@ let vueDeleteUser = new Vue({
       $('#modal-admin-users-delete').removeClass('is-active')
     },
     deleteUser: (ev) => {
-      vueDeleteUser.cancel()
+      vueDeleteUser.loading = true
+      $.ajax('/admin/users/' + usrData._id, {
+        dataType: 'json',
+        method: 'DELETE'
+      }).then((rData, rStatus, rXHR) => {
+        vueDeleteUser.loading = false
+        vueDeleteUser.cancel()
+        window.location.assign('/admin/users')
+      }, (rXHR, rStatus, err) => {
+        vueDeleteUser.loading = false
+        alerts.pushError('Error', rXHR.responseJSON.msg)
+      })
     }
   }
 })